refactor: atom을 각기 다른 관련 있는 모듈에서 관리하도록 수정
