+++ src/components/common/ServerSelect/ServerSelect.tsx
@@ -1,10 +1,10 @@
 import { ChangeEventHandler, useEffect, useState } from 'react';
 import { useRecoilState } from 'recoil';
 import styled from 'styled-components';
-import { serverOriginState } from '../../../recoil/atoms';
 import { BASE_URLS } from '../../../constants/api';
 import { isKeyOf } from '../../../types/typeGuards';
 import useCartService from '../../../hooks/useCartService';
+import { serverOriginState } from '../../../recoil/atoms/common';
 
 const ServerSelect = () => {
   const [value, setValue] = useState('baron');
