+++ src/recoil/atoms/cart.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import type { CartItem } from '../../types/cart';
+
+export const cartState = atom<CartItem[]>({
+  key: 'cartState',
+  default: [],
+});
